Refactor the test in the package org. Apache. Dubbo. Reactive.
I observed and found the responseObserver mock. In ManyToManyMethodHandlerTest,

ManyToOneMethodHandlerTest and OneToManyMethodHandlerTest these three have repeatedly creating problems in the test class.

To avoid duplicate creation. I created a class called creatObserverAdapter and set the get method. In the test cases of the three test classes, the get method is used to avoid the repeated creation of mocks and stubs.